// client/src/setup/store/actionsTypes/technologyActionType.ts
export enum TechnologiesActionType {
  REQUEST_GET_TECHNOLOGIES = "REQUEST_GET_TECHNOLOGIES ",
  SUCCESS_GET_TECHNOLOGIES = "SUCCESS_GET_TECHNOLOGIES",
  FAILED_GET_TECHNOLOGIES = "FAILED_GET_TECHNOLOGIES",
  REQUEST_CREATE_TECHNOLOGY = "REQUEST_CREATE_TECHNOLOGY ",
  SUCCESS_CREATE_TECHNOLOGY = "SUCCESS_CREATE_TECHNOLOGY",
  FAILED_CREATE_TECHNOLOGY = "FAILED_CREATE_TECHNOLOGY",
  REQUEST_UPDATE_TECHNOLOGY = "REQUEST_UPDATE_TECHNOLOGY ",
  SUCCESS_UPDATE_TECHNOLOGY = "SUCCESS_UPDATE_TECHNOLOGY",
  FAILED_UPDATE_TECHNOLOGY = "FAILED_UPDATE_TECHNOLOGY",
  REQUEST_DELETE_TECHNOLOGY = "REQUEST_DELETE_TECHNOLOGY ",
  SUCCESS_DELETE_TECHNOLOGY = "SUCCESS_DELETE_TECHNOLOGY",
  FAILED_DELETE_TECHNOLOGY = "FAILED_DELETE_TECHNOLOGY",
}